Move locale and command info into core settings
