+++ src/main/kotlin/org/cascadebot/cascadebot/data/objects/GuildData.kt
@@ -28,19 +28,17 @@ class GuildData(@field:Id val guildId: Long) {
     private val creationDate = Date()
     //endregion
 
-    private val commandInfo = ConcurrentHashMap<Class<ICommandMain>, MutableSet<GuildCommandInfo>>()
-    val guildCommandInfos
-        get() = Collections.unmodifiableMap(commandInfo)
-
     val enabledFlags: MutableSet<Flag> = Sets.newConcurrentHashSet()
 
-    val locale: Locale = Locale.getDefaultLocale()
+    // TODO: Keep this here or remove it? It just serves as an accessor for the actual locale
+    val locale: Locale
+        get() = coreSettings.locale
 
     //region Guild data containers
     val coreSettings = GuildSettingsCore()
     val usefulSettings = GuildSettingsUseful()
     val moderationSettings = GuildSettingsModeration()
-    val permissionSettings = GuildPermissions()
+    val managementSettings = GuildSettingsManagement()
     val musicSettings = GuildSettingsMusic()
     //endregion
 
@@ -74,68 +72,6 @@ class GuildData(@field:Id val guildId: Long) {
         }
     }
 
-    //region Commands
-    fun enableCommand(command: ICommandMain) {
-        if (command.module.isPrivate) return
-        if (commandInfo.contains(command.javaClass)) {
-            getGuildCommandInfo(command).enabled = true
-        }
-    }
-
-    fun enableCommandByModule(module: Module) {
-        if (module.isPrivate) return
-        for (command in CascadeBot.INS.commandManager.getCommandsByModule(module)) {
-            enableCommand(command)
-        }
-    }
-
-    fun disableCommand(command: ICommandMain) {
-        if (command.module.isPrivate) return
-        getGuildCommandInfo(command).enabled = false
-    }
-
-    fun disableCommandByModule(module: Module) {
-        if (module.isPrivate) return
-        for (command in CascadeBot.INS.commandManager.getCommandsByModule(module)) {
-            disableCommand(command)
-        }
-    }
-
-    fun isCommandEnabled(command: ICommandMain): Boolean {
-        return if (commandInfo.contains(command.javaClass)) {
-            getGuildCommandInfo(command).enabled
-        } else coreSettings.isModuleEnabled(command.module)
-    }
-
-    fun getCommandAliases(command: ICommandMain): Set<String> {
-        return if (commandInfo.contains(command.javaClass)) {
-            getGuildCommandInfo(command).aliases.applyChanges(command.getGlobalAliases(locale))
-        } else command.getGlobalAliases(locale)
-    }
-
-    fun addAlias(command: ICommandMain, alias: String): Boolean {
-        return getGuildCommandInfo(command).aliases.add(alias)
-    }
-
-    fun removeAlias(command: ICommandMain, alias: String): Boolean {
-        return getGuildCommandInfo(command).aliases.remove(alias)
-    }
-
-    private fun getGuildCommandInfoSet(command: ICommandMain): MutableSet<GuildCommandInfo> {
-        return commandInfo.computeIfAbsent(command.javaClass) {
-            mutableSetOf(GuildCommandInfo(command, locale))
-        }
-    }
-
-    private fun getGuildCommandInfo(command: ICommandMain, locale: Locale = this.locale): GuildCommandInfo {
-        return getGuildCommandInfoSet(command).find { it.locale == locale } ?: run {
-            val commandInfo = GuildCommandInfo(command, locale)
-            getGuildCommandInfoSet(command).add(commandInfo)
-            commandInfo
-        }
-    }
-
-    //endregion
     fun enableFlag(flag: Flag): Boolean {
         return enabledFlags.add(flag)
     }
