+++ src/main/kotlin/org/cascadebot/cascadebot/data/objects/GuildSettingsCore.kt
@@ -1,8 +1,11 @@
 package org.cascadebot.cascadebot.data.objects
 
 import com.google.common.collect.Sets
+import org.cascadebot.cascadebot.CascadeBot
+import org.cascadebot.cascadebot.commandmeta.ICommandMain
 import org.cascadebot.cascadebot.commandmeta.Module
 import org.cascadebot.cascadebot.data.Config
+import org.cascadebot.cascadebot.data.language.Locale
 import java.util.concurrent.ConcurrentHashMap
 
 @SettingsContainer(module = Module.CORE)
@@ -35,6 +38,11 @@ class GuildSettingsCore {
     @Setting
     var helpShowAllModules = false
 
+    var locale: Locale = Locale.getDefaultLocale()
+
+    @Setting(directlyEditable = false)
+    private val commandInfo = ConcurrentHashMap<Class<ICommandMain>, MutableSet<GuildCommandInfo>>()
+
     @Setting(directlyEditable = false)
     var prefix: String = Config.INS.defaultPrefix
 
@@ -64,6 +72,69 @@ class GuildSettingsCore {
             return true
         }
         return isEnabled
+    }
+
+    //endregion
+
+    //region Commands
+    fun enableCommand(command: ICommandMain) {
+        if (command.module.isPrivate) return
+        if (commandInfo.contains(command.javaClass)) {
+            getGuildCommandInfo(command).enabled = true
+        }
+    }
+
+    fun enableCommandByModule(module: Module) {
+        if (module.isPrivate) return
+        for (command in CascadeBot.INS.commandManager.getCommandsByModule(module)) {
+            enableCommand(command)
+        }
+    }
+
+    fun disableCommand(command: ICommandMain) {
+        if (command.module.isPrivate) return
+        getGuildCommandInfo(command).enabled = false
+    }
+
+    fun disableCommandByModule(module: Module) {
+        if (module.isPrivate) return
+        for (command in CascadeBot.INS.commandManager.getCommandsByModule(module)) {
+            disableCommand(command)
+        }
+    }
+
+    fun isCommandEnabled(command: ICommandMain): Boolean {
+        return if (commandInfo.contains(command.javaClass)) {
+            getGuildCommandInfo(command).enabled
+        } else isModuleEnabled(command.module)
+    }
+
+    fun getCommandAliases(command: ICommandMain): Set<String> {
+        return if (commandInfo.contains(command.javaClass)) {
+            getGuildCommandInfo(command).aliases.applyChanges(command.getGlobalAliases(locale))
+        } else command.getGlobalAliases(locale)
+    }
+
+    fun addAlias(command: ICommandMain, alias: String): Boolean {
+        return getGuildCommandInfo(command).aliases.add(alias)
+    }
+
+    fun removeAlias(command: ICommandMain, alias: String): Boolean {
+        return getGuildCommandInfo(command).aliases.remove(alias)
+    }
+
+    private fun getGuildCommandInfoSet(command: ICommandMain): MutableSet<GuildCommandInfo> {
+        return commandInfo.computeIfAbsent(command.javaClass) {
+            mutableSetOf(GuildCommandInfo(command, locale))
+        }
+    }
+
+    private fun getGuildCommandInfo(command: ICommandMain, locale: Locale = this.locale): GuildCommandInfo {
+        return getGuildCommandInfoSet(command).find { it.locale == locale } ?: run {
+            val commandInfo = GuildCommandInfo(command, locale)
+            getGuildCommandInfoSet(command).add(commandInfo)
+            commandInfo
+        }
     }
 
     //endregion
